+++ youtube_client_app/src/app/app.module.ts
@@ -4,7 +4,7 @@ import AppRoutingModule from './app-routing.module';
 import AppComponent from './app.component';
 import AuthModule from './auth/auth.module';
 import CoreModule from './core/core.module';
-import { MainPageModule } from './main-page/main-page.module';
+import { YoutubeModule } from './youtube/youtube.module';
 import { Page404Module } from './page404/page404.module';
 
 @NgModule({
@@ -14,7 +14,7 @@ import { Page404Module } from './page404/page404.module';
     AppRoutingModule,
     AuthModule,
     CoreModule,
-    MainPageModule,
+    YoutubeModule,
     Page404Module,
   ],
   bootstrap: [AppComponent],
