+++ youtube_client_app/src/app/core/models/search-item.model.ts
