--- youtube_client_app/src/app/main-page/pages/main-page.component.html
@@ -1,1 +1,0 @@
-<app-search-list></app-search-list>
