--- youtube_client_app/src/app/main-page/search/search-list/search-list.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { SearchService } from '../../services/search.service';
-
-@Component({
-  selector: 'app-search-list',
-  templateUrl: './search-list.component.html',
-  styleUrls: ['./search-list.component.scss'],
-})
-export default class SearchListComponent {
-  public constructor(protected searchService: SearchService) {}
-}
