+++ youtube_client_app/src/app/youtube/components/search-results/search-item/search-item.component.scss
