+++ youtube_client_app/src/app/youtube/components/search-results/search-results.component.scss
