+++ youtube_client_app/src/app/youtube/components/search-results/search-results.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+import { SearchService } from 'src/app/core/services/search.service';
+
+@Component({
+  selector: 'app-search-results',
+  templateUrl: './search-results.component.html',
+  styleUrls: ['./search-results.component.scss'],
+})
+export default class SearchResultsComponent {
+  public constructor(protected searchService: SearchService) {}
+}
