+++ youtube_client_app/src/app/youtube/pages/main-page/main-page.component.html
@@ -1,0 +1,1 @@
+<app-search-results></app-search-results>
